section1/requirements.txt
pandas==1.5.3
openpyxl==3.1.2
psycopg2-binary==2.9.5